+++ psi4/src/psi4/libscf_solver/hf.h
@@ -312,7 +312,7 @@ class HF : public Wavefunction {
     virtual double compute_E();
 
     /** Applies second-order convergence acceleration */
-    virtual int soscf_update(float soscf_conv, int soscf_min_iter, int soscf_max_iter, int soscf_print);
+    virtual int soscf_update(double soscf_conv, int soscf_min_iter, int soscf_max_iter, int soscf_print);
 
     /// Figure out how to occupy the orbitals in the absence of DOCC and SOCC
     void find_occupation();
