+++ psi4/src/psi4/libscf_solver/rohf.h
@@ -87,7 +87,7 @@ class ROHF : public HF {
     void finalize();
 
     void damping_update(double);
-    int soscf_update(float soscf_conv, int soscf_min_iter, int soscf_max_iter, int soscf_print);
+    int soscf_update(double soscf_conv, int soscf_min_iter, int soscf_max_iter, int soscf_print);
     bool stability_analysis();
 
     std::shared_ptr<ROHF> c1_deep_copy(std::shared_ptr<BasisSet> basis);
