+++ psi4/src/psi4/libscf_solver/uhf.h
@@ -83,7 +83,7 @@ class UHF : public HF {
     void finalize();
 
     void damping_update(double);
-    int soscf_update(float soscf_conv, int soscf_min_iter, int soscf_max_iter, int soscf_print);
+    int soscf_update(double soscf_conv, int soscf_min_iter, int soscf_max_iter, int soscf_print);
     bool stability_analysis();
 
     /// Hessian-vector computers and solvers
